refactor: prepare library structure for Godot
