+++ .gitignore
@@ -1,0 +1,8 @@
+# Build artifacts
+*.exe
+*.o
+/build/
+
+# VSCode files
+.vscode/
+
