+++ include/errorHandler.h
@@ -1,0 +1,12 @@
+#ifndef ERROR_HANDLER_H
+#define ERROR_HANDLER_H
+
+#include <iostream>
+#include <stdexcept>
+
+class inputErrors {
+public:
+    bool playerAmountError(int playerAmount);
+};
+
+#endif
